requirements.txt
facebook-sdk==0.4.0
requests==2.9.1
wheel==0.26.0
